Provide functionality for extension to swap between before and after pages. (#9)
The extension now provides two buttons that swap the pages and swap them back again respectively.

Also, the extension is now only enabled on sites on our test site domain (well also some other pages like new tab page, seems like it will take more unnecessary work to exclude these). I thought this might be a good idea because right now our extension is quite limited in terms of the cases it can work with, and our model may work well only with some sites to begin with, which we can easily add for.

Also, I adjusted the layout of the UI for the Pop up of the extension.